Moved all window related settings to specific file and created function to call in dllmain in place of old system. Updated window handler to use new settings properly. Fullscreen mode now uses the same window as borderless to fix Alt-Tab issue, also locks mouse to window.
