+++ SADXModLoader/window.h
@@ -1,0 +1,5 @@
+#pragma once
+
+extern HINSTANCE g_hinstDll;
+
+void PatchWindow(const LoaderSettings& settings, wstring borderpath);
